BUG: Fix parsing attributes from TDS 5.0 dataset.xml
The inclusion of the NCML namespace was breaking the parsing because
ElementTree hands us tag names with the namespace included. Fix by
stripping out the namespace, which seems to be the most common solution.
Also take the opportunity to move handler lookup to _Types so that we
don't make an external reference to _Types.__dict__ (Ew.)