Add fixtures for IGRA2 HTTP tests
This relies on a VCRpy hook to modify the http response before writing
to disk. This allows us to drastically reduce the data volume in the
cassette, which is not only good to reduce disk size from ~200MB, but
also makes the test run 100x faster.

Also remove no-longer used zip files from FTP-based tests.