Recalculate RAOB from Wyoming
This change allows for a user to request that the RAOB is
"Recalculated." This can be required if an earlier user of the
site requests the observation before the entire observation is
available. This results in an incomplete observation being cached.
Recalculation requires additional resources, so the default is to
not recalculate.