BUG: Fix incorrect URLs in hyrax catalog
THREDDS Client Catalog spec explicitly says that the service base and
the urlpath should be concatenated directly. So avoid using urljoin(),
which was treating the urlpath as an absolute path on the server. This
breaks a weird link in a ramadda catalog. Just adjust the test for
ramadda and treat its urlpath (with full server) as a bug.